Module udt::UdtOpts

Various options that can be passed to getsockopt or setsockopt

These are typed in such a way so that when they are used with getsockopt or setsockopt, they will require the right data type.

Examples

use udt::*;

let sock = UdtSocket::new(SocketFamily::AFInet, SocketType::Stream).unwrap();
let recv_buf: i32 = sock.getsockopt(UdtOpts::UDT_RCVBUF).unwrap();
let rendezvous: bool = sock.getsockopt(UdtOpts::UDT_RENDEZVOUS).unwrap();

Structs

UDP_RCVBUF
UDP socket receiver buffer size (bytes)

UDP_SNDBUF
UDP socket sender buffer size (bytes)

UDT_EVENT
The EPOLL events available to this socket. Read only.

UDT_FC
Maximum window size (packets)

UDT_LINGER
Linger time on close().

UDT_MAXBW
Maximum bandwidth that one single UDT connection can use (bytes per second).

UDT_MSS
Maximum Packet size (bytes)

UDT_RCVBUF
UDT receiver buffer size limit (bytes)

UDT_RCVDATA
Size of data available to read, in the receiving buffer. Read only.

UDT_RCVSYN
Synchronization mode for receiving.

UDT_RCVTIMEO
Receiving call timeout (milliseconds).

UDT_RENDEZVOUS
Rendezvous connection setup.

UDT_REUSEADDR
Reuse an existing address or create a new one.

UDT_SNDBUF
UDT sender buffer size limit (bytes)

UDT_SNDDATA
Size of pending data in the sending buffer. Read only.

UDT_SNDSYN
Synchronization mode of data sending

UDT_SNDTIMEO
Sending call timeout (milliseconds).

UDT_STATE
Current status of the UDT socket. Read only.